Create Find The Spot
A text based game based on a Cartesian Plane system that I developed in high school. Subject may wander around in an attempt to find treasure. I added obstacles at various points in the game to prevent the user from reaching the end of the game. For example, there are warp holes, pits etc. I also built in a loophole where I can use a password (cheat code of sorts) to log in as administrator and warp myself to whatever co-ordinate I want.