Create Sum of Multiple Elements (Array Implementation)
This program demonstrates the limitations and advantages of an array. Having been statically declared, the size is fixed, and cannot exceed 100 numbers. If the number of elements exceeds 100, the program will not be of any use. In this case, a dynamically array, or a vector would be more appropriate. These are used for list of elements where the size of the list is not known.