Create Binary Search
Use of for loops to implement the binary search technique. Initial assumption is that element that needs to be found is in the center of the array. Then depending on whether elements required is greater than or less than the element that was found, the position can be changed to (middle-1 + start)/2 or (middle+1 + end)/2.